Refactor long division
Long division is fast but at the expense of a decrease in performance
of multiplication. To rectify that we can remove the constraint set
by prefix3, i.e. that base must be as large as 2^17, so that
multiplication remains as performant as it was before.

To remove the prefix3 constraint we have to do multi-digit division
of r3/d2 in order to compute the first estimation. We use slowDivModBy
to do that. This allows us to start back using a base as large as 2^26.

The end result is that we have a long division algorithm with reasonable
performance without sacrificing the performance of our multiplication.