Compare Natural against BigInt
Natural beat BigInt on all benchmarks besides compare.

I learned that BigInt is using a much better comparison algorithm than
I came up with. Basically, we don't need to reverse or compute lengths
to compare when the digits are stored in little-endian format. There's
a clever algorithm that works for the little-endian format.